Prediction method for the HAL SuperLearner wrapper

method for predict for SL.hal

Quick pass-through.

# S3 method for SL.hal
predict(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)

# S3 method for SL.hal
predict(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)

Arguments

object

A fitted object of class hal

newdata

A matrix of new predictions on which to obtain predictions

bigDesign

A boolean indicating whether to obtain predictions all at once (which may be memory intensive) or to split up the task into smaller chunks

chunks

A numeric indicating how many observations to use in each chunk of the prediction task (if bigDesign = FALSE)

...

Other arguments passed to predict