Highly Adaptive Lasso

Model fitting function.

hal(Y, X, newX = NULL, family = gaussian(), verbose = F,
  obsWeights = rep(1, length(Y)), sparseMat = TRUE,
  nfolds = ifelse(length(Y) <= 100, 20, 10), nlambda = 100, useMin = TRUE,
  debug = T, parallel = F, ...)

Arguments

Y

outcome

X

data

newX

New data to apply the model fit and generate predictions.

family

Statistical family

verbose

Set to T for more detailed output

obsWeights

observation weights

sparseMat

Use sparse matrix implementation or normal matrix implementation. Normal matrix implementation is old and not guaranteed to work correctly.

nfolds

Number of CV folds for cv.glmnet

nlambda

Number of lambda values to test in cv.glmnet.

useMin

Glmnet option - use minimum risk lambda or 1se lambda (more penalization).

debug

Setting to T will run garbage collection to improve the accuracy of memory monitoring.

parallel

Use a regisered parallel backend if possible.